Update to kinect2_tracker.hpp: publish depth frame
There wasn't a published rostopic for the Kinect V2 depth frame.  These changes publish the depth frame array as a 16 bit grayscale image which is converted from the 1 mm depth pixel format defined in the OniEnums.h file (OpenNI2).  The rostopic is: /kinect_depth, which stores the raw depth data.  This is particularly useful for manipulating raw data from the Kinect V2 and storing depth images.